Simplify secrets management
Instead of having a variable with the current secret name and a list
with previous names, have a single list and use the last item as current
while marking all others as 'to be removed'.
Simplifies changing the variable, as one only has to add a single line
with next number in sequence.