Make the runs idempotent
When the secret backend is referenced via alias, kv-v2, hashivault treats
it almost correctly, but insists that it has changed things on each run.

Using a more standard way of passing version as an option, this is not
the case and a playbook run correctly reports that nothing was changed.